bug fix
数据格式问题造成的解析问题fix

+++ YiYao/DTO/MTMHealthCollectDTO.cs
@@ -8,15 +8,15 @@ namespace YiYao.DTO
 {
     class MTMHealthCollectDTO
     {
-        public int eatingpreference { get; set; }
+        public double eatingpreference { get; set; }
         public string smoking { get; set; }
         public string drinking { get; set; }
-        public int height { get; set; }
-        public int weight { get; set; }
-        public int waist { get; set; }
+        public double height { get; set; }
+        public double weight { get; set; }
+        public double waist { get; set; }
         public int heartrate { get; set; }
-        public int systolicpressure { get; set; }
-        public int diastolicpressure { get; set; }
+        public double systolicpressure { get; set; }
+        public double diastolicpressure { get; set; }
         public float fastsugar { get; set; }
         public float randomsugar { get; set; }
         public float HbA1c { get; set; }
